Move non-cluster tests to ConsumerGroupCommandTest
